+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 
 
 ## [5.0.3] - 2021-01-07
+### Changed
+- mysql_class.Server.upd_srv_stat:  Replaced percentage equations with calls to gen_libs.pct_int.
 
 
 ## [5.0.2] - 2020-11-16