+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 
 ## [5.0.3] - 2021-01-07
 ### Changed
+- mysql_libs.create_instance:  Added "rep_user" and "rep_japd" to be allowed to be passed to class instance if detected.
 - mysql_class.SlaveRep.\_\_init\_\_:  Added self.rep_user and self.rep_japd for replication user information.
 - mysql_class.Server.connect:  Set any error connection messages to self.conn_msg.
 - mysql_class.Server.connect:  Added "silent" keyword argument to prevent printing any connection error messages.